soa-betting-app/requirements.txt
fastapi[all]
uvicorn
alembic
sqlalchemy
dependency_injector
containers
pydantic
python-dotenv
psycopg2
requests
